So many fixes

+++ fs.py
@@ -1,18 +1,14 @@
 import pickle
 from os.path import join
 from os import listdir
+from os import path
+import re
 
-def upgradeWad():
-    x = None
-    with open("data.wad", "rb") as f:
-        x = f.read()
-        x = pickle.loads(x)
-        x = wad(x)
-    
-    with open("data.wad", "wb") as f:
-        x = pickle.dumps(x)
-        f.write(x)
-
+def checkWad():
+    if not path.exists('data.wad'):
+        open('data.wad', 'x')
+        with open('data.wad', 'wb') as f:
+            pickle.dump(wad(), f)
 class wad:
     def __init__(self, oldWad = {}):
         self.bookLines = oldWad
@@ -44,10 +40,8 @@ def listBooks():
 
 def getBook(fileName):
     with open(join("books",fileName), "rt") as f:
-        v = f.readlines()
-        v = [x for x in v if x != "\n"]
-        v = [x for x in v if x != ""]
-        v = [x for x in v if x != "****"]
+        v = f.read()
+        v = re.split(r' *[\.\?!][\'"\)\]]* *', v)
         return v
 
 def readLine(book):
